backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-cors-headers==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
importlib-resources==6.0.1
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
pkgutil_resolve_name==1.3.10
pytz==2023.3
referencing==0.30.2
requests==2.31.0
rpds-py==0.9.2
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
zipp==3.16.2
gunicorn==20.1.0
